requirements.txt
streamlit
phidata
groq
duckduckgo-search
google-generativeai